Move the keyframe transform calculation to Animation.h
