+++ Sprocket/CMakeLists.txt
@@ -59,6 +59,7 @@ add_library(Sprocket STATIC
             Scene/Systems/GameGrid.cpp
             Scene/Systems/PathFollower.cpp
 
+            Graphics/Animation.cpp
             Graphics/Buffer.cpp
             Graphics/BufferLayout.cpp
             Graphics/RenderContext.cpp
